Separate schema in individual files
